requirements.txt
xarray
netcdf4
requests
pyarrow
pandas
psycopg2-binary
tenacity
apache-airflow